requirements.txt
tensorflow==2.9.1
keras==2.9.0
keract==4.5.0
umap-learn==0.5.3
